97-move-out-geometries: spread geometries structs into individual files.
